updating project to run locally

+++ src/main/java/com/rene/ecommerce/services/ClientService.java
@@ -7,20 +7,17 @@ import java.util.Optional;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.rene.ecommerce.domain.Product;
+import com.rene.ecommerce.domain.dto.updated.UpdatedClient;
 import com.rene.ecommerce.domain.users.Client;
 import com.rene.ecommerce.exceptions.AuthorizationException;
 import com.rene.ecommerce.exceptions.ClientOrSellerHasThisSameEntryException;
 import com.rene.ecommerce.exceptions.DuplicateEntryException;
 import com.rene.ecommerce.exceptions.ObjectNotFoundException;
-import com.rene.ecommerce.exceptions.ProductHasAlreadyBeenSold;
-import com.rene.ecommerce.exceptions.YouHaveAlreadyAddThisProductInYourWishlistException;
+import com.rene.ecommerce.exceptions.UserHasProductsRelationshipsException;
 import com.rene.ecommerce.repositories.ClientRepository;
-import com.rene.ecommerce.repositories.ProductRepository;
 import com.rene.ecommerce.repositories.SellerRepository;
 import com.rene.ecommerce.security.ClientSS;
 
@@ -32,17 +29,12 @@ public class ClientService {
 
 	@Autowired
 	private SellerRepository sellerRepo;
-	
-	@Autowired
-	private ProductRepository productRepo;
+
 
 
 	@Autowired
 	private BCryptPasswordEncoder passwordEncoder;
 	
-
-	@Autowired
-	private ProductService productService;
 
 
 	public Client findById(Integer id) {
@@ -85,18 +77,25 @@ public class ClientService {
 	}
 
 	@Transactional
-	public Client update(Client obj) {
+	public Client update(UpdatedClient obj) {
 		ClientSS user = UserService.clientAuthenticated();
-		obj.setPassword(passwordEncoder.encode(obj.getPassword()));
+		
+		Client cli = findById(user.getId());
 
-
-		if (user == null || !user.getId().equals(obj.getId())) {
+		if (user == null || !user.getId().equals(cli.getId())) {
 			throw new AuthorizationException();
 		}
 		
-		if (sellerRepo.findByEmail(obj.getEmail()) == null) {
+		cli.setCpf(obj.getCpf());
+		cli.setEmail(obj.getEmail());
+		cli.setName(obj.getName());
+		cli.setPassword(passwordEncoder.encode(obj.getPassword()));
+
+
+		
+		if (sellerRepo.findByEmail(cli.getEmail()) == null) {
 			try {
-				return clientRepo.save(obj);
+				return clientRepo.save(cli);
 			} catch (Exception e) {
 				throw new DuplicateEntryException();
 			}
@@ -108,35 +107,22 @@ public class ClientService {
 
 	public void delete() {
 		ClientSS user = UserService.clientAuthenticated();
-
-		try {
+		
+		Client cli = findById(user.getId());
+		
+		// verify if the client hasn't bought any products
+		// doing this by numberOfBuys because the performance
+		if(cli.getNumberOfBuys() == 0) {
 			clientRepo.deleteById(user.getId());
-		} catch (DataIntegrityViolationException e) {
-			throw new DataIntegrityViolationException("You can't delete this object");
 		}
+		
+		else {
+			throw new UserHasProductsRelationshipsException();
+
+		}
+		
 	}
 	
-	public void setProductAsWished(Integer productId) {
-		
-		Product product = productService.findById(productId);
-		
-		if(Product.isSold(product)) {
-			throw new ProductHasAlreadyBeenSold();
-		}
-		
-		ClientSS user = UserService.clientAuthenticated();
-		Client client = findById(user.getId());
-		
-		if(client.getProductsWished().contains(product)) {
-			throw new YouHaveAlreadyAddThisProductInYourWishlistException();
-		}
-		
 
-		client.getProductsWished().add(product);
-		product.getWhoWhishesThisProduct().add(client);
-		
-		clientRepo.save(client);
-		productRepo.save(product);
-	}
 
 }
